Add Prettier to Rollup
Added Prettier to Rollup bundle generation.

Made watch and build generate two diferent files instead of the same as the file in watch mode should not be minified due to better debug errors.

Disable typescript option
oEmitOnError only work in build mode as it exits the process early when in watch mode.